Ensure FeatureState changes are not automaticaly "persisted" in InMemoryStateRepository
Prevents scenarios where a `FeatureState` is set into `InMemoryStateRepository`
but is changed after without an explicit call to
`InMemoryStateRepository.setFeatureState`. In thoses cases, we still
want to return an object with the state that was stored, not the current
changed one, to be consistent with other `StateRepositories`, like
`JDBCStateRepository`.

We always store and return copies of `FeatureState` in `InMemoryStateRepository`
to prevent mutability of those states.