Reload features asynchronously when Executor provided
Assume the following use case: If a service serves, for example, 10k
req/s and every request checks 10 feature flags with a TTL of 60s,
access to the remote cache takes 10ms. When a user passes a custom
ExecutorService to the CachingStateRepository, features will be updated
asynchronously. The client thread won't be blocked; it receives either
null when a feature is unavailable in the cache, the value when it is
available and not expired, or an expired value and scheduled job
reloads it asynchronously.

Returning null is not issue as DefaultFeatureManager replaces it with
default value.

If client doesn't pass Executor CachingStateRepository will work with
old blocking approach.